requirements.txt
attrs==16.1.0
py==1.4.31
pytest==3.0.1
